Add CNAME file to preserve custom domain
Currently, the custom domain is being wiped from github on each push. This is because configuring a custom domain in github creates a CNAME record. However, our deploy system force pushes, which nukes that record. As such, we need to include it explicitly. With this CNAME record in place, the custom domain should persist.