Replace Google Analytics v4 with Universal Analytics
Google recently released a new version of analytics. This requires a code change to support. Until we make such a change (better to do so across all reports) we should stick to the older Universal Analytics version. This is a new measurement ID for a new GA property for the report, using that (older) implementation.